Extract dispersement and loaning logic to dedicated utility module
