+++ src/app/shared/models/goal.model.ts
@@ -1,5 +1,6 @@
 import { Budget } from './budget.model';
 import { setMembership } from '../utils/membership';
+import { disperseBalance } from '../utils/dispersement';
 
 export enum GoalStatus {
   Funded,
@@ -173,7 +174,7 @@ export class Goal implements IGoal {
     if (this.current > this.target) {
       const difference = this.current - this.target;
       this.current -= difference;
-      this.budget.disperse(difference);
+      disperseBalance(this.budget, difference);
     }
   }
 
@@ -226,26 +227,5 @@ export class Goal implements IGoal {
     } else {
       return GoalStatus.Normal;
     }
-  }
-
-  public canLoanTo(recipient: Goal): boolean {
-    if (recipient == this) {
-      // A goal can always "loan" to itself
-      return true;
-    }
-    if (this.isFunded()) {
-      // A funded goal has no capacity to loan to other goals
-      return false
-    }
-    if (this.isPurchased()) {
-      // Purchased goals have a negative balance and cannot loan to any other goal
-      return false;
-    }
-    if (this.isEarmarked()) {
-      // Earmarked goals can only loan to cover balance of already purchased goals
-      return recipient.isPurchased();
-    }
-    // If any of the other situations don't exist, then it's safe for this goal to loan to any goal
-    return true;
   }
 }
